\encoding{UTF-8}
\name{abmiweb}
\alias{abmiweb}
\title{Open ABMI Website}
\description{
The function opens the website of the Alberta Biodiversity Monitoring Institute (ABMI) in the default navigator.
Checkout for news and data stes.
}
\usage{
abmiweb()
}
\references{
\url{http://www.abmi.ca}
}
\author{
\enc{Péter Sólymos}{Peter Solymos}, \email{solymos@ualberta.ca}
}
\examples{
\dontrun{
abmiweb()
}
}
\keyword{ documentation }
