\encoding{UTF-8}
\name{abmi-package}
\alias{abmi-package}
\alias{abmi}
\docType{package}
\title{
Data Processing and Analysis for the Alberta Biodiversity Monitoring Institute
}
\description{
Data Processing and Analysis for the Alberta Biodiversity Monitoring Institute.
Data processing, biodiversity intactness calculations and simulations.

To be continued.
}
\author{
\enc{P\'eter S\'olymos}{Peter Solymos}, \email{solymos@ualberta.ca}
}
\references{
\url{http://www.abmi.ca/}
}
\keyword{ package }
\examples{
\dontrun{
abmiweb()
}
}
